Finished Login Page
Finished login page for the most part. Added functionality to the hide password buttons. Have not yet added functionality to the login buttons which redirect to another page that contains a typing field for the user to input their disability. No need for a continue with google button. Modify color gradient as needed.